Moved EditVenue to Features/Venues
