+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/DeleteVenue/DeleteVenue.cshtml
@@ -162,7 +162,7 @@
             var confirm = $('#Confirm');
 
             function disableDeleteBtn() {
-                $('pttcd-delete-location__delete-btn')
+                $('#pttcd-delete-location__delete-btn')
                     .attr('disabled', confirm.is(':checked') ? null : 'disabled');
             }
 
@@ -170,4 +170,4 @@
             confirm.change(disableDeleteBtn);
         })(jQuery);
     </script>
-} 
+}
