+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/EditVenue/Address.cshtml
@@ -1,4 +1,4 @@
-﻿@model Dfc.CourseDirectory.WebV2.Features.EditVenue.Address.Command
+﻿@model Dfc.CourseDirectory.WebV2.Features.Venues.EditVenue.Address.Command
 @{
     ViewContext.SetLayoutData(title: "Location address");
 }
