+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/EditVenue/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model Dfc.CourseDirectory.WebV2.Features.EditVenue.Details.ViewModel
+﻿@model Dfc.CourseDirectory.WebV2.Features.Venues.EditVenue.Details.ViewModel
 @{
     ViewContext.SetLayoutData(title: "Check your location details before saving");
 }
