+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/EditVenue/Email.cshtml
@@ -1,4 +1,4 @@
-﻿@model Dfc.CourseDirectory.WebV2.Features.EditVenue.Email.Command
+﻿@model Dfc.CourseDirectory.WebV2.Features.Venues.EditVenue.Email.Command
 @{
     ViewContext.SetLayoutData(title: "Email address");
 }
