+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/EditVenue/Name.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Dfc.CourseDirectory.Core.Validation.VenueValidation
-@model Dfc.CourseDirectory.WebV2.Features.EditVenue.Name.Command
+@model Dfc.CourseDirectory.WebV2.Features.Venues.EditVenue.Name.Command
 @{
     ViewContext.SetLayoutData(title: "Location name");
 }
