+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/EditVenue/PhoneNumber.cshtml
@@ -1,4 +1,4 @@
-﻿@model Dfc.CourseDirectory.WebV2.Features.EditVenue.PhoneNumber.Command
+﻿@model Dfc.CourseDirectory.WebV2.Features.Venues.EditVenue.PhoneNumber.Command
 @{
     ViewContext.SetLayoutData(title: "Phone number");
 }
