+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/EditVenue/Website.cs
@@ -9,7 +9,7 @@ using MediatR;
 using OneOf;
 using OneOf.Types;
 
-namespace Dfc.CourseDirectory.WebV2.Features.EditVenue.Website
+namespace Dfc.CourseDirectory.WebV2.Features.Venues.EditVenue.Website
 {
     public class Query : IRequest<Command>
     {
