+++ src/Dfc.CourseDirectory.WebV2/Features/Venues/EditVenue/Website.cshtml
@@ -1,4 +1,4 @@
-﻿@model Dfc.CourseDirectory.WebV2.Features.EditVenue.Website.Command
+﻿@model Dfc.CourseDirectory.WebV2.Features.Venues.EditVenue.Website.Command
 @{
     ViewContext.SetLayoutData(title: "Website");
 }
